requirements-dev.txt
-r requirements.txt
black
mypy
django-stubs[compatible-mypy]
ruff
isort
pylint-django
types-Pygments